+++ src/Kit/DataDisplay/IdCard/index.tsx
@@ -5,22 +5,23 @@ import {FunctionComponent} from 'react';
 import {useKitTheme} from '@theme/useKitTheme';
 import {IKitIdCard} from './types';
 import {KitTypography} from '@kit/General';
+import KitAvatar from '../Avatar/Avatar';
 
-const _getTitle = (title: IKitIdCard['title'], cardSize: IKitIdCard['size']) => {
+const _getTitle = (title: IKitIdCard['title'], cardSize: IKitIdCard['size'], disabled: boolean) => {
     const clx = cn({'kit-id-card-title-large': cardSize === 'large'});
 
     return (
-        <KitTypography.Text className={clx} size="large" weight="bold" ellipsis={{tooltip: true}}>
+        <KitTypography.Text className={clx} size="large" weight="bold" ellipsis={{tooltip: true}} disabled={disabled}>
             {title}
         </KitTypography.Text>
     );
 };
 
-const _getDescription = (description: IKitIdCard['description'], cardSize: IKitIdCard['size']) => {
+const _getDescription = (description: IKitIdCard['description'], cardSize: IKitIdCard['size'], disabled: boolean) => {
     const textSize = cardSize === 'medium' ? 'small' : 'large';
 
     return (
-        <KitTypography.Text size={textSize} ellipsis={{tooltip: true}}>
+        <KitTypography.Text size={textSize} ellipsis={{tooltip: true}} disabled={disabled}>
             {description}
         </KitTypography.Text>
     );
@@ -30,26 +31,31 @@ export const KitIdCard: FunctionComponent<IKitIdCard> = ({
     className,
     title,
     description,
-    avatar,
+    avatarProps,
     size = 'medium',
+    disabled = false,
     ...idCardProps
 }) => {
     const {appId} = useKitTheme();
 
     const clx = cn(appId, styles['kit-id-card'], className);
 
-    const shouldDisplayAvatar = avatar !== undefined;
+    const shouldDisplayAvatar = avatarProps !== undefined;
     const shouldDisplayTitle = title !== undefined;
     const shouldDisplayDescription = description !== undefined;
     const shouldDisplayInfo = shouldDisplayTitle || shouldDisplayDescription;
 
     return (
         <div className={`${clx} id-card-grid`} {...idCardProps}>
-            {shouldDisplayAvatar && <div className="card-avatar">{avatar}</div>}
+            {shouldDisplayAvatar && (
+                <div className="card-avatar">
+                    <KitAvatar {...avatarProps} disabled={disabled} />
+                </div>
+            )}
             {shouldDisplayInfo && (
                 <div className="card-info">
-                    {shouldDisplayTitle && _getTitle(title, size)}
-                    {shouldDisplayDescription && _getDescription(description, size)}
+                    {shouldDisplayTitle && _getTitle(title, size, disabled)}
+                    {shouldDisplayDescription && _getDescription(description, size, disabled)}
                 </div>
             )}
         </div>