+++ stories/02-Design System/05-DataDisplay/IdCard/examples/basic.tsx
@@ -1,5 +1,5 @@
-import {KitIdCard} from '@kit/DataDisplay';
 import React from 'react';
+import {KitIdCard} from '@kit/DataDisplay';
 
 const App = () => {
     return <KitIdCard title="John Doe" description="A talented software developer." />;