+++ stories/02-Design System/05-DataDisplay/IdCard/test-components/LargeIdCard.tsx
@@ -1,13 +1,19 @@
 import React, {FunctionComponent} from 'react';
 import {KitTypography} from '@kit/General';
-import {KitIdCard, KitAvatar} from '@kit/DataDisplay';
+import {KitIdCard} from '@kit/DataDisplay';
+import {IKitAvatar} from '@kit/DataDisplay/Avatar/types';
+
+const avatar: IKitAvatar = {
+    shape: 'square',
+    children: 'JD'
+};
 
 export const LargeIdCardTest: FunctionComponent = () => {
     return (
         <div>
             <KitTypography.Title level="h3">Large size IdCard</KitTypography.Title>
             <KitIdCard
-                avatar={<KitAvatar shape="square">JD</KitAvatar>}
+                avatarProps={avatar}
                 title="John Doe"
                 description="A talented software developer."
                 size="large"