+++ stories/02-Design System/05-DataDisplay/ItemList/examples/actions.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import {KitSpace} from '@kit/Layout';
-import {KitAvatar, KitItemList} from '@kit/DataDisplay';
+import {KitItemList} from '@kit/DataDisplay';
 import {faCartPlus, faStar, faBullhorn} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
+import {IKitIdCard} from '@kit/DataDisplay/IdCard/types';
 
 const App = () => {
-    const idCardProps = {
+    const idCardProps: IKitIdCard = {
         title: 'Tondeuse à gazon',
         description: 'Tondeuse thermique Auto tractée 70 VL 55 TH',
-        avatar: <KitAvatar label={'TG'} />
+        avatarProps: {label: 'TG'}
     };
 
     const [action1, action2, action3] = [