+++ stories/02-Design System/05-DataDisplay/ItemList/examples/basic.tsx
@@ -1,18 +1,19 @@
 import React from 'react';
 import {KitSpace} from '@kit/Layout';
-import {KitAvatar, KitItemList} from '@kit/DataDisplay';
+import {KitItemList} from '@kit/DataDisplay';
+import {IKitAvatar} from '@kit/DataDisplay/Avatar/types';
 
 const App = () => {
     const title = 'Tondeuse à gazon';
     const description = 'Tondeuse thermique Auto tractée 70 VL 55 TH';
-    const avatar = <KitAvatar label={'TG'} />;
+    const avatarProps: IKitAvatar = {label: 'TG'};
 
     return (
         <KitSpace direction="vertical" size="m">
             <div style={{width: '550px', display: 'flex', flexDirection: 'column', gap: '10px'}}>
                 <KitItemList idCardProps={{title}} />
                 <KitItemList idCardProps={{title, description}} />
-                <KitItemList idCardProps={{title, description, avatar}} />
+                <KitItemList idCardProps={{title, description, avatarProps}} />
             </div>
         </KitSpace>
     );