+++ stories/02-Design System/05-DataDisplay/ItemList/examples/content.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import {KitSpace} from '@kit/Layout';
-import {KitAvatar, KitItemList} from '@kit/DataDisplay';
+import {KitItemList} from '@kit/DataDisplay';
+import {IKitIdCard} from '@kit/DataDisplay/IdCard/types';
 
 const App = () => {
-    const idCardProps = {
+    const idCardProps: IKitIdCard = {
         title: 'Tondeuse à gazon',
         description: 'Tondeuse thermique Auto tractée 70 VL 55 TH',
-        avatar: <KitAvatar label={'TG'} />
+        avatarProps: {label: 'TG'}
     };
 
     const baseStyle = {