+++ stories/02-Design System/05-DataDisplay/ItemList/examples/draggable.tsx
@@ -5,9 +5,10 @@ import {CSS} from '@dnd-kit/utilities';
 import {arrayMove, SortableContext, useSortable, verticalListSortingStrategy} from '@dnd-kit/sortable';
 import {faDeleteLeft, faGripLines} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {KitAvatar, KitItemList} from '@kit/DataDisplay';
+import {KitItemList} from '@kit/DataDisplay';
 import {KitButton} from '@kit/General';
 import {KitSpace} from '@kit/Layout';
+import {IKitIdCard} from '@kit/DataDisplay/IdCard/types';
 
 type Item = {
     id: number;
@@ -15,10 +16,10 @@ type Item = {
     actions: ComponentProps<typeof KitItemList>['actions'];
 };
 
-const idCardProps = {
+const idCardProps: IKitIdCard = {
     title: 'Tondeuse à gazon',
     description: 'Tondeuse thermique Auto tractée 70 VL 55 TH',
-    avatar: <KitAvatar label={'TG'} />
+    avatarProps: {label: 'TG'}
 };
 
 const actions = [